fix(go): rename internal pkg to shadow
