+++ Cargo.toml
@@ -5,7 +5,7 @@ authors = ["clearloop <udtrokia@gmail.com>"]
 edition = "2018"
 description = "The shadow service for relayers and verify workers to retrieve header data and generate proof. Shadow will index the data it needs from blockchain nodes, such as Ethereum and Darwinia."
 repository = "https://github.com/darwinia-network/shadow"
-license = "GPL"
+license = "GPL-3.0-or-later"
 documentation = "https://docs.rs/darwinia-shadow"
 homepage = "https://github.com/darwinia-network/shadow"
 include = [
