+++ pkg/shadow/eth/types.go
