+++ pkg/shadow/util/collection.go
