+++ pkg/shadow/util/encode.go
